chore: refactoring, update README.md
