--- revscoring_model/model_servers/__init__.py
@@ -1,8 +1,0 @@
-from .model_server_mp import RevscoringModelMP
-from .model_servers import RevscoringModel, RevscoringModelType
-
-__all__ = [
-    "RevscoringModel",
-    "RevscoringModelType",
-    "RevscoringModelMP",
-]
